requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.58.5
greenlet==3.1.1
h11==0.14.0
holidays==0.76
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.2.6
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.3.0
prophet==1.1.7
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-calamine==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stanio==0.5.1
starlette==0.46.1
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
xlrd==2.0.2
xlsxwriter==3.2.5
